Refactor AgendaDbContext seeder to use ServiceProvider
Refactored the Seeder() method in the AgendaDbContext class to use the ServiceProvider initialized in App.xaml.cs instead of creating a new usermanager and only being given the context.

Further commented out the original Console/Program.cs file as it no longer would work with the new changes to database seeding.

Small adjustments to App.xaml.cs to deal with nullability related warnings.